+++ hypervisor/Makefile
@@ -198,7 +198,7 @@ ifeq ($(CONFIG_MULTIBOOT2),y)
 BOOT_C_SRCS += boot/multiboot/multiboot2.c
 endif
 BOOT_C_SRCS += boot/reloc.c
-BOOT_C_SRCS += arch/x86/ptcm.c
+BOOT_C_SRCS += arch/x86/rtcm.c
 
 # hardware management component
 HW_S_SRCS += arch/x86/idt.S
